Removed notebooks from initial PR and updated stein_2d_toy.py to new run signature.
