+++ docs/source/index.rst
@@ -57,6 +57,8 @@ NumPyro documentation
    examples/gp
    examples/bnn
    examples/stein_bnn
+   examples/stein_mixture_dmm
+   examples/stein_2d_toy
    examples/dais_demo
    examples/sparse_regression
    examples/horseshoe_regression
