+++ notebooks/source/index.rst
@@ -17,8 +17,6 @@ Welcome to NumPyro Tutorials!
    bayesian_hierarchical_linear_regression
    discrete_imputation
    model_rendering
-   stein_kernels
-   stein_neutra
 
 
 Indices
