+++ test/test_examples.py
@@ -44,8 +44,9 @@ EXAMPLES = [
     "prodlda.py --num-steps 10 --hidden 10 --nn-framework haiku",
     "sparse_regression.py --num-samples 10 --num-warmup 10 --num-data 10 --num-dimensions 10",
     "ssbvm_mixture.py --num-samples 10 --num-warmup 10",
+    "stein_2d_toy.py --max-iter 10",
     "stein_bnn.py --max-iter 10 --subsample-size 10 --num-particles 5",
-    "stein_mixture_dmm.py --max-iter 5 --subsample-size 77",
+    "stein_mixture_dmm.py --max-iter 2 --subsample-size 2 --testing true --num-particles 1",
     "stochastic_volatility.py --num-samples 100 --num-warmup 100",
     "ucbadmit.py --num-chains 2",
     "vae.py -n 1",
